backend/requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn>=0.20.0,<0.23.0
pydantic>=1.10.0,<2.0.0
sqlalchemy>=2.0.0,<2.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
slack-sdk>=3.20.0
python-multipart>=0.0.5
langchain>=0.0.200
openai>=0.27.0
numpy>=1.20.0
